+++ server/tsuserver.py
@@ -68,8 +68,8 @@ class TsuserverDR:
         self.release = 4
         self.major_version = 3
         self.minor_version = 2
-        self.segment_version = 'a8'
-        self.internal_version = 'm220521a'
+        self.segment_version = 'a9'
+        self.internal_version = 'm220521b'
         version_string = self.get_version_string()
         self.software = 'TsuserverDR {}'.format(version_string)
         self.version = 'TsuserverDR {} ({})'.format(version_string, self.internal_version)